# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Alpogo API Documentation",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/account/login": {
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "Logea un usuario",
                "description": "",
                "operationId": "loginUser",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "The user email for login",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "The password for login in clear text",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/account/registration": {
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "Registra un usuario",
                "description": "",
                "operationId": "post_user",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "201": {
                        "description": "Create successfully"
                    }
                }
            }
        },
...
        "/event/": {
            "post": {
                "tags": [
                    "event"
                ],
                "summary": "Crea un evento",
                "description": "",
                "operationId": "add_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            }
        },
        "/event/{slug}": {
            "get": {
                "tags": [
                    "event"
                ],
                "summary": "Detalle de un evento",
                "description": "",
                "operationId": "show_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "201": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "event"
                ],
                "summary": "Edita un evento",
                "description": "",
                "operationId": "update_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "408": {
                        "description": "Event not found"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "event"
                ],
                "summary": "Edita parcialmente un evento",
                "description": "",
                "operationId": "partial_update_event",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "408": {
                        "description": "Event not found"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/event/{slug}/items": {
            "get": {
                "tags": [
                    "event"
                ],
                "summary": "Lista de items de un evento",
                "description": "Retorna la lista de items de un eventos",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "408": {
                        "description": "Event not found"
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Event": {
            "required": [
                "title",
                "short_description",
                "description",
                "cover_image",
                "start_date",
                "end_date",
                "age_restiction",
                "locale",
                "location"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "example": "Jhon"
                },
                "short_description": {
                    "type": "string",
                    "example": "Lorem ipsum"
                },
                "description": {
                    "type": "string",
                    "example": "Lorem ipsum"
                },
                "cover_image": {
                    "type": "string",
                    "example": ""
                },
                "start_date": {
                    "type": "string",
                    "example": "1989/09/04"
                },
                "end_date": {
                    "type": "string",
                    "example": "1989/09/04"
                },
                "age_restriction": {
                    "type": "string",
                    "example": "18"
                },
                "locale": {
                    "type": "string",
                    "example": "Club Paraguay"
                },
                "location": {
                    "type": "string",
                    "example": "C\u00f3rdoba, Argentina"
                }
            },
            "type": "object",
            "xml": {
                "name": "Event"
            }
        },
...
        "User": {
            "required": [
                "first_name",
                "last_name",
                "email",
                "avatar",
                "passport",
                "birthday"
            ],
            "properties": {
                "first_name": {
                    "type": "string",
                    "example": "Jhon"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "email": {
                    "type": "string",
                    "example": "jhon@doe.com"
                },
                "avatar": {
                    "type": "string",
                    "example": "/jhon/doe/avatar.jpg"
                },
                "passport": {
                    "type": "string",
                    "example": "34335174"
                },
                "birthday": {
                    "type": "string",
                    "example": "1989/04/09"
                },
                "password": {
                    "type": "string",
                    "example": "jhondoe"
                },
                "password_confirmation": {
                    "type": "string",
                    "example": "jhondoe"
                }
            },
            "type": "object",
            "xml": {
                "name": "User"
            }
        }
    }
}